refactor: reorganize sample project packages
- com.example.invoker/ — Calculator (@Invoker demos)
- com.example.constructor/ — Point (@Constructor on class)
- com.example.record/ — Coord (@Constructor on record)

Each annotation type gets its own package, avoiding Constructor
caller class collisions and making the demos clearer.
Update version to 0.0.5.
