+++ samples/maven-sample/src/main/java/com/example/constructor/Point.java
@@ -1,4 +1,4 @@
-package com.example.rawit;
+package com.example.constructor;
 
 import rawit.Constructor;
